Amna's latest calc
This new version has all functionalities sqroot,tan,cos,sin,log2,log10
and _ + * /

The way I did this was more like how the mac calculator does it.
Like if you enter
3+2+---> (at this point the screen says 6+)
3*6+----->(the screen says 6 because it knows its not supposed to add
the 6 right now)

This is a smarter algorithm BUT if the user does not use it properly,it
won't give the right results.

I liked the google calculator as well. I have another idea of how to
implement this. The other idea would be more user friendly.
Like the screen would display the chain of calculations done by the user
so its going to be less error prone but at the same time it will be less
smarter with the predictions as the user will be able to delete old
numbers too.
Like 3+2+5 . the user will be able to delete the 2 .

Right now you all can try these calculations on your calculator and on
the calculator app.
I have the right results.
(the answers I mentioned are rounded of)
Match your real calculator results to this.